Add an OpenGL renderer
It seems like SDL has no API to draw a triangle. So use OpenGL API
to render to texture and then copy (glGetTexImage) the results to
an SDL_Texture.

At the beginning, it keeps running into segment fault as long as writing
to SDL_Texture::pixels. It turns out that it is critical restriction
in SDL that an SDL_Texture is not lockable unless it is created with
SDL_TEXTUREACCESS_STREAMING access mode.

Signed-off-by: Luc Ma <luc@sietium.com>